Changed the way color buffers are stored in the class
