+++ TongFang/Enums.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Text;
 
 namespace TongFang
@@ -155,208 +154,5 @@ namespace TongFang
         NUMPAD_PERIOD,
         PRINT_SCREEN,
         BACKSLASH
-    }
-
-    public static class Layouts
-    {
-        public static readonly Dictionary<Key, byte> ISO = new Dictionary<Key, byte>
-        {
-            { Key.LEFT_CONTROL , 105},
-            { Key.FN_Key , 107},
-            { Key.LEFT_WINDOWS , 108},
-            { Key.LEFT_ALT , 109},
-            { Key.SPACE , 112},
-            { Key.RIGHT_ALT , 115},
-            { Key.MENU , 116},
-            { Key.RIGHT_CONTROL , 117},
-            { Key.ARROW_LEFT , 118},
-            { Key.DOWN_ARROW , 119},
-            { Key.RIGHT_ARROW , 120},
-            { Key.NUMPAD_0 , 121},
-            { Key.NUMPAD_PERIOD , 122},
-            { Key.LEFT_SHIFT , 84},
-            { Key.BACKSLASH_UK , 86},
-            { Key.Z , 87},
-            { Key.X , 88},
-            { Key.C , 89},
-            { Key.V , 90},
-            { Key.B , 91},
-            { Key.N , 92},
-            { Key.M , 93},
-            { Key.COMMA , 94},
-            { Key.PERIOD , 95},
-            { Key.FORWARD_SLASH , 96},
-            { Key.RIGHT_SHIFT , 97},
-            { Key.UP_ARROW , 98},
-            { Key.NUMPAD_ONE , 99},
-            { Key.NUMPAD_TWO , 100},
-            { Key.NUMPAD_THREE , 101},
-            { Key.NUMPAD_ENTER , 102},
-            { Key.CAPS_LOCK , 63},
-            { Key.A , 65},
-            { Key.S , 66},
-            { Key.D , 67},
-            { Key.F , 68},
-            { Key.G , 69},
-            { Key.H , 70},
-            { Key.J , 71},
-            { Key.K , 72},
-            { Key.L , 73},
-            { Key.SEMICOLON , 74},
-            { Key.APOSTROPHE , 75},
-            { Key.HASHTAG , 76},
-            { Key.NUMPAD_FOUR , 78},
-            { Key.NUMPAD_FIVE , 79},
-            { Key.NUMPAD_SIX , 80},
-            { Key.TAB , 42},
-            { Key.Q , 44},
-            { Key.W , 45},
-            { Key.E , 46},
-            { Key.R , 47},
-            { Key.T , 48},
-            { Key.Y , 49},
-            { Key.U , 50},
-            { Key.I , 51},
-            { Key.O , 52},
-            { Key.P , 53},
-            { Key.OPEN_BRACKET , 54},
-            { Key.CLOSE_BRACKET , 55},
-            { Key.ENTER , 56},
-            { Key.NUMPAD_SEVEN , 57},
-            { Key.NUMPAD_EIGHT , 58},
-            { Key.NUMPAD_NINE , 59},
-            { Key.NUMPAD_PLUS , 60},
-            { Key.TILDE , 21},
-            { Key.ONE , 22},
-            { Key.TWO , 23},
-            { Key.THREE , 24},
-            { Key.FOUR , 25},
-            { Key.FIVE , 26},
-            { Key.SIX , 27},
-            { Key.SEVEN , 28},
-            { Key.EIGHT , 29},
-            { Key.NINE , 30},
-            { Key.ZERO , 31},
-            { Key.MINUS , 32},
-            { Key.EQUALS , 33},
-            { Key.BACKSPACE , 35},
-            { Key.NUMLOCK , 36},
-            { Key.NUMPAD_DIVIDE , 37},
-            { Key.NUMPAD_MULTIPLY , 38},
-            { Key.NUMPAD_MINUS , 39},
-            { Key.ESCAPE , 0},
-            { Key.F1 , 1},
-            { Key.F2 , 2},
-            { Key.F3 , 3},
-            { Key.F4 , 4},
-            { Key.F5 , 5},
-            { Key.F6 , 6},
-            { Key.F7 , 7},
-            { Key.F8 , 8},
-            { Key.F9 , 9},
-            { Key.F10 , 10},
-            { Key.F11 , 11},
-            { Key.F12 , 12},
-            { Key.INSERT , 13},
-            { Key.DELETE , 14},
-            { Key.HOME , 15},
-            { Key.END , 16},
-            { Key.PAGE_UP , 17},
-            { Key.PAGE_DOWN , 18 }
-        };
-
-        public static readonly Dictionary<Key, byte> ANSI = new Dictionary<Key, byte>
-        {
-            { Key.ESCAPE , 0},
-            { Key.F1 , 1},
-            { Key.F2 , 2},
-            { Key.F3 , 3},
-            { Key.F4 , 4},
-            { Key.F5 , 5},
-            { Key.F6 , 6},
-            { Key.F7 , 7},
-            { Key.F8 , 8},
-            { Key.F9 , 9},
-            { Key.F10 , 10},
-            { Key.F11 , 11},
-            { Key.F12 , 12},
-            { Key.INSERT , 13},
-            { Key.PRINT_SCREEN , 14},
-            { Key.DELETE , 15 },
-
-            { Key.TILDE , 21},
-            { Key.ONE , 22},
-            { Key.TWO , 23},
-            { Key.THREE , 24},
-            { Key.FOUR , 25},
-            { Key.FIVE , 26},
-            { Key.SIX , 27},
-            { Key.SEVEN , 28},
-            { Key.EIGHT , 29},
-            { Key.NINE , 30},
-            { Key.ZERO , 31},
-            { Key.MINUS , 32},
-            { Key.EQUALS , 33},
-            { Key.BACKSPACE , 35},
-            { Key.HOME , 36},
-
-            { Key.TAB , 42},
-            { Key.Q , 44},
-            { Key.W , 45},
-            { Key.E , 46},
-            { Key.R , 47},
-            { Key.T , 48},
-            { Key.Y , 49},
-            { Key.U , 50},
-            { Key.I , 51},
-            { Key.O , 52},
-            { Key.P , 53},
-            { Key.OPEN_BRACKET , 54},
-            { Key.CLOSE_BRACKET , 55},
-            { Key.BACKSLASH , 56},
-            { Key.PAGE_UP , 57},
-
-            { Key.CAPS_LOCK , 63},
-            { Key.A , 65},
-            { Key.S , 66},
-            { Key.D , 67},
-            { Key.F , 68},
-            { Key.G , 69},
-            { Key.H , 70},
-            { Key.J , 71},
-            { Key.K , 72},
-            { Key.L , 73},
-            { Key.SEMICOLON , 74},
-            { Key.APOSTROPHE , 75},
-            { Key.ENTER , 77},
-            { Key.PAGE_DOWN , 78},
-
-            { Key.LEFT_SHIFT , 85},
-            { Key.Z , 87},
-            { Key.X , 88},
-            { Key.C , 89},
-            { Key.V , 90},
-            { Key.B , 91},
-            { Key.N , 92},
-            { Key.M , 93},
-            { Key.COMMA , 94},
-            { Key.PERIOD , 95},
-            { Key.FORWARD_SLASH , 96},
-            { Key.RIGHT_SHIFT , 97},
-            { Key.UP_ARROW , 98},
-            { Key.END , 99},
-
-            { Key.LEFT_CONTROL , 105},
-            { Key.FN_Key , 106},
-            { Key.LEFT_WINDOWS , 108},
-            { Key.LEFT_ALT , 109},
-            { Key.SPACE , 112},
-            { Key.RIGHT_ALT , 115},
-            { Key.MENU , 116},
-            { Key.RIGHT_CONTROL , 117},
-            { Key.ARROW_LEFT , 118},
-            { Key.DOWN_ARROW , 119},
-            { Key.RIGHT_ARROW , 120},
-        };
     }
 }
